removed dangerous bulk-password-functions.js, consolidated to 1 file, added awareness of quotes per feedback
