+++ scripts/bulk-password-update-export.js
@@ -2,7 +2,9 @@
 process.env.NODE_TLS_REJECT_UNAUTHORIZED=0;
 const minimist = require('minimist');
 const {promises: fsPromises} = require('fs');
-const { areYouSure, changeUserPass, generatePassword } = require('./bulk-password-functions.js');
+const readline = require('readline');
+const rpn = require('request-promise-native');
+const {randomInt} = require('crypto');
 
 const argv = minimist(process.argv.slice(2), {
   alias: {
@@ -15,7 +17,7 @@ if (argv.h|| !argv.url || !argv.user || !argv.password || !argv.use_passes) {
 
 Usage:
       node bulk-password-update-export.js -h | --help
-      node bulk-password-update-export.js --url https://localhost --user medic --password adminPass
+      node bulk-password-update-export.js --url https://localhost --user medic --password adminPass --use_passes false
 
 Options:
     -h --help     Show this screen.
@@ -49,6 +51,70 @@ const options = {
   }
 };
 
+const areYouSure = async (userCount, extraWarning = '') => {
+  const NOTE = ` ****** WARNING ******\
+
+Continuing will log ${userCount} users out of the CHT until you provide them with their updated password.
+${extraWarning}
+Do you want to continue? [y/N]
+`;
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+
+  return new Promise(resolve => rl.question(NOTE, ans => {
+    rl.close();
+    resolve(ans);
+  }));
+};
+
+const changeUserPass = async (user, password, options) => {
+  const admins = ['admin', 'medic'];
+  const adminError = new Error('403 - Cannot change password for "medic" or "admin" users.');
+  const postOptions = {...options};
+  postOptions.body = {
+    'password': password
+  };
+  postOptions.uri = `${options.uri}/${user}`;
+  try {
+    if (admins.includes(user)) {
+      throw adminError;
+    }
+    await rpn.post(postOptions);
+    console.log('SUCCESS', user, password);
+  } catch (e) {
+    console.log('ERROR', user, e.message);
+  }
+};
+
+const generatePassword = async () => {
+  const CHAR_COUNT = 4;
+
+  const CHARS = 'ABCDEFGHJKLMNPQRSTUVWXYZ';
+  const LOWER_CHARS = CHARS.toLowerCase();
+  const NUMS = '0123456789';
+  const rando = (chars) => {
+    return Array(CHAR_COUNT)
+        .fill('')
+        .map(() => chars.charAt(randomInt(chars.length - 1)))
+        .join('');
+  };
+
+  // thanks https://stackoverflow.com/a/3943985 & https://stackoverflow.com/a/6274381
+  const randoShuffle = function (orderedString) {
+    const unorderedArray = orderedString.split('');
+    const stringLength = (orderedString.length - 1);
+    for (let i = stringLength; i > 0; i--) {
+      const j = Math.floor(randomInt(2) * (i + 1));
+      [unorderedArray[i], unorderedArray[j]] = [unorderedArray[j], unorderedArray[i]];
+    }
+    return unorderedArray.join('');
+  };
+  // CHT requires 8 minimum so we'll do 4 upper, 4 lower, 4 int
+  return randoShuffle(rando(CHARS) + rando(LOWER_CHARS) + '-' + rando(NUMS));
+};
+
 // thanks https://bobbyhadz.com/blog/javascript-read-file-into-array !
 const loadUsers = async () => {
   try {
@@ -61,14 +127,30 @@ const loadUsers = async () => {
   }
 };
 
+
+// Thanks https://stackoverflow.com/a/40672956
+String.prototype.splitCSV = function() {
+  let matches = this.match(/(\s*"[^"]+"\s*|\s*[^,]+|,)(?=,|$)/g);
+  for (let n = 0; n < matches.length; ++n) {
+    matches[n] = matches[n].trim();
+    if (matches[n] === ',') matches[n] = '';
+  }
+  return matches;
+}
+
 const getUserAndPassword = async (user) => {
   let newPass;
   let newUser = user;
   if ( argv.use_passes === 'true') {
-    const user_array = user.toString().split(',');
+    const user_array = user.splitCSV();
     if ( user_array[0] && user_array[1] ){
       newUser = user_array[0].toString().trim();
-      newPass = user_array[1].toString().trim();
+      // regexp removes quotes only from first and last chars
+      // So "wFB38p,GM" will be wFB38p,GM
+      newPass = user_array[1]
+          .toString()
+          .trim()
+          .replace(/^"(.*)"$/, '$1');
     }
   } else {
     newPass = await generatePassword();
